requirements.txt
SQLAlchemy>=2.0.23
aiosqlite>=0.19.0
playwright>=1.40.0
httpx>=0.25.2
python-telegram-bot>=20.7
beautifulsoup4>=4.12.0
playwright-stealth>=1.0.6
